Refactor framework Catalog entities to utilize TCS columns
Add Map where create name/value pairs

Add List where entity doesn't exist

+++ src-framework/org/broadleafcommerce/order/domain/PaymentInfo.java
@@ -1,30 +1,32 @@
 package org.broadleafcommerce.order.domain;
 
+import java.math.BigDecimal;
+
 import org.broadleafcommerce.profile.domain.Address;
 
 public interface PaymentInfo {
 
-//	public Object getPayment();
-//	
-//	public void setPayment(Object payment);
-	
-	public Long getId();
-	
-	public void setId(Long id);
-	
-	public Order getOrder();
-	
-	public void setOrder(Order order);
-	
-	public Address getAddress();
-	
-	public void setAddress(Address address);
-	
-	public double getAmount();
-	
-	public void setAmount(double amount);
-	
-	public String getReferenceNumber();
-	
-	public void setReferenceNumber(String referenceNumber);
+    //	public Object getPayment();
+    //
+    //	public void setPayment(Object payment);
+
+    public Long getId();
+
+    public void setId(Long id);
+
+    public Order getOrder();
+
+    public void setOrder(Order order);
+
+    public Address getAddress();
+
+    public void setAddress(Address address);
+
+    public BigDecimal getAmount();
+
+    public void setAmount(BigDecimal amount);
+
+    public String getReferenceNumber();
+
+    public void setReferenceNumber(String referenceNumber);
 }
